Use docker for developing and running SyPet
Added a Dockerfile that permits creating a docker container set up with a
development environment for SyPet. It is also possible to run SyPet from docker.

Added a Makefile that permits doing that without needing to type large commands.

The README.md was also updated with information about how to build and run
SyPet, with or without docker.