Fix classpath, build file
The jar for sat4j must appear before apt on the classpath because apt already
includes a version of sat4j (that we do not want).

Cleaned up the build file a bit, removing some extra stuff, and adding back an
option for creating a jar file, which was removed unintendally.