Clean up SyPet, SyPetInput, SyPetConfig and JsonParser
Overall, the code is cleaner. This is work in progress. The code compiles, but
it is not working yet.

SyPet has been refactored to allow for better testing. Also, the idea is that,
in the future, SyPet might be called with other kinds of input, besides JSON
files (for example, options in the command line).

SyPetInput and SyPetConfig now apply the builder pattern (refer to Effective
Java, 3rd edition, Item 2, for an overview of this pattern). This gives a better
API for constructing objects with many fields.

JsonParser was also cleaned up a fair bit, and defers setting default parameters
to class constructors.